release 25.4.1

+++ src/main/java/jpize/util/NativeLib.java
@@ -12,6 +12,12 @@ public class NativeLib {
 
     public static void load(Resource res) {
         try{
+            if(getOsName().equals("android")) {
+                final String libName = res.simpleName().replace("lib", "");
+                System.loadLibrary(libName);
+                return;
+            }
+
             final InputStream stream = res.inStream();
             final TempFileResource temp = Resource.temp(res.simpleName(), res.extension());
             temp.deleteOnExit();
@@ -22,11 +28,25 @@ public class NativeLib {
         }
     }
 
+    public static String getOsName() {
+        final String os = System.getProperty("os.name").toLowerCase();
+        if(os.contains("win"))
+            return "windows";
+
+        if(os.contains("linux")) {
+            final String runtimeName = System.getProperty("java.runtime.name", "").toLowerCase();
+            if(runtimeName.contains("android"))
+                return "android";
+            return "linux";
+        }
+        return os;
+    }
+
     public static String getFilename(String libname) {
-        final String osName = System.getProperty("os.name").toLowerCase();
-        if(osName.contains("win"))
+        final String os = getOsName();
+        if(os.contains("win"))
             return (libname + ".dll");
-        if(osName.contains("mac"))
+        if(os.contains("mac"))
             return ("lib" + libname + ".dylib");
         return ("lib" + libname + ".so");
     }
